added mock functionality to categories

+++ api/src/categories/categories.controller.ts
@@ -1,10 +1,13 @@
 import { Controller, Get, Post, Body, Patch, Param, Delete, NotFoundException } from '@nestjs/common';
+import { DeleteResult } from 'typeorm';
 import { CategoriesService } from './categories.service';
 import { CreateCategoryDto } from './dto/create-category.dto';
 import { UpdateCategoryDto } from './dto/update-category.dto';
 import { Mockable } from '../shared/mockable.decorator';
 import { MockableController } from '../shared/mockable.controller';
 import { MOCK_CATEGORIES } from './mock/data.mock';
+import { MockCategory } from './mock/category.mock';
+import { Pretty } from '../types/utility';
 
 @Controller('categories')
 export class CategoriesController extends MockableController {
@@ -12,11 +15,59 @@ export class CategoriesController extends MockableController {
     super();
   }
 
+  private handleMockCreate(createCategoryDto: CreateCategoryDto) {
+    const nextIndex = MOCK_CATEGORIES.length;
+
+    const newCategory = new MockCategory({ name: createCategoryDto.name, index: nextIndex });
+
+    return this.simulateResponse(() => newCategory);
+  }
+
+  private handleMockGetMany() {
+    return this.simulateResponse<typeof MOCK_CATEGORIES>(() => MOCK_CATEGORIES);
+  }
+
+  private async handleMockGetOne(id: number) {
+    const found = await this.simulateResponse<typeof MOCK_CATEGORIES[number] | null>(() => {
+      return MOCK_CATEGORIES.find((category) => category.id === id) ?? null;
+    });
+
+    if (!found) {
+      throw new NotFoundException();
+    }
+
+    return found;
+  }
+
+  private async handleMockUpdate(id: number, updateCategoryDto: UpdateCategoryDto) {
+    const found = await this.handleMockGetOne(id);
+
+    Object.assign(found, updateCategoryDto);
+
+    return found;
+  }
+
+  private async handleMockDelete(id: number) {
+    await this.handleMockGetOne(id);
+
+    type Res = Pretty<DeleteResult>;
+
+    return {
+      raw: '',
+      affected: 1
+    } satisfies Res;
+  }
+
+
   @Post()
   create(
     @Mockable() mock: boolean,
     @Body() createCategoryDto: CreateCategoryDto
   ) {
+    if (mock) {
+      return this.handleMockCreate(createCategoryDto);
+    }
+
     return this.categoriesService.create(createCategoryDto);
   }
 
@@ -25,27 +76,19 @@ export class CategoriesController extends MockableController {
     @Mockable() mock: boolean
   ) {
     if (mock) {
-      return this.simulateResponse<typeof MOCK_CATEGORIES>(() => MOCK_CATEGORIES);
+      return this.handleMockGetMany();
     }
 
     return this.categoriesService.findAll();
   }
 
   @Get(':id')
-  async findOne(
+  findOne(
     @Mockable() mock: boolean,
     @Param('id') id: string
   ) {
     if (mock) {
-      const found = await this.simulateResponse<typeof MOCK_CATEGORIES[number] | null>(() => {
-        return MOCK_CATEGORIES.find((category) => category.id === +id) ?? null;
-      });
-
-      if (!found) {
-        throw new NotFoundException();
-      }
-
-      return found;
+      return this.handleMockGetOne(+id);
     }
 
     return this.categoriesService.findOne(+id);
@@ -57,6 +100,10 @@ export class CategoriesController extends MockableController {
     @Param('id') id: string,
     @Body() updateCategoryDto: UpdateCategoryDto
   ) {
+    if (mock) {
+      return this.handleMockUpdate(+id, updateCategoryDto);
+    }
+
     return this.categoriesService.update(+id, updateCategoryDto);
   }
 
@@ -65,6 +112,10 @@ export class CategoriesController extends MockableController {
     @Mockable() mock: boolean,
     @Param('id') id: string
   ) {
+    if (mock) {
+      return this.handleMockDelete(+id);
+    }
+
     return this.categoriesService.remove(+id);
   }
 }
